feat(front-kit-3): moved newslatterLayout
