+++ packages/falcon-ui-kit/src/Newsletter/index.ts
@@ -1,0 +1,1 @@
+export * from './NewsletterLayout';
