+++ packages/falcon-ui-kit/src/index.ts
@@ -18,6 +18,7 @@ export * from './Locale';
 export * from './Logo';
 export * from './Menu';
 export * from './MiniCart';
+export * from './Newsletter';
 export * from './Order';
 export * from './Price';
 export * from './Product';
